Using templates
As it is a requirement to do most of the work on the client side, I
can’t properly use Go’s templates. So for each DB table there’ll be a
web page where the user can do the CRUD ops via ajax calls. As I can’t
find a way to do dynamic template selection (that is: {{template
dinamicNameHere}}) nor I can’t spend too much time with it, I’ll
duplicate the html for each table (Categorias, Noticias, etc.) and each
one will be a template on its own with a route to each one.